+++ tests/test_search_request.py
@@ -1,9 +1,20 @@
+import pytest
+from pydantic import ValidationError
+
 from scim2_models.rfc7644.search_request import SearchRequest
 
 
 def test_search_request():
     SearchRequest(
         attributes=["userName", "displayName"],
+        filter='userName Eq "john"',
+        sort_by="userName",
+        sort_order=SearchRequest.SortOrder.ascending,
+        start_index=1,
+        count=10,
+    )
+
+    SearchRequest(
         excluded_attributes=["timezone", "phoneNumbers"],
         filter='userName Eq "john"',
         sort_by="userName",
@@ -42,3 +53,24 @@ def test_count_floor():
 
     sr = SearchRequest(count=-1)
     assert sr.count == 1
+
+
+def test_attributes_or_excluded_attributes():
+    """Test that a validation error is raised when both 'attributes' and 'excludedAttributes' are filled at the same time.
+    https://datatracker.ietf.org/doc/html/rfc7644#section-3.9
+
+        Clients MAY request a partial resource representation on any
+        operation that returns a resource within the response by specifying
+        either of the mutually exclusive URL query parameters "attributes" or
+        "excludedAttributes"...
+    """
+
+    payload = {
+        "schemas": ["urn:ietf:params:scim:api:messages:2.0:SearchRequest"],
+        "attributes": ["userName"],
+        "excludedAttributes": [
+            "displayName",
+        ],
+    }
+    with pytest.raises(ValidationError):
+        SearchRequest.model_validate(payload)